pyproject.toml
[project]
name = "qbenchsim-dataset"
version = "0.1.0"
description = "QBenchSim Dataset"
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "customtkinter>=5.2.2",
    "ipykernel>=6.30.0",
    "mqt-bench>=1.1.9",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "quantum-executor>=0.1.0",
    "sacred>=0.8.7",
    "setuptools>=80.9.0",
]
